Refactor: useTxSender
